refactor: shift all contest routes to dedicated contest router
